barman keep: Make barman-cloud WAL cleanup keep-aware
Makes the code which handles cleanup of WALs when barman-cloud deletes
a backup aware of archival backups.

This re-uses the new `BackupManager.should_remove_wals` method added
to determine whether WALs should be removed and which WAL ranges should
be protected. The logic in `_remove_wals_for_backup` is then updated to
check that no WALs are within the protected ranges.

As with the WAL cleanup logic in BackupManager, we also add a
`skip_wal_cleanup_if_standalone` option to the `_delete_backup` and
`_remove_wals_for_backup` methods so that we only clean up WALs in the
case where all older backups are standalone archival backups if the
deletion is the result of applying a retention policy.

This is part of the implementation of #370.