Remove userepl from has_backup_privileges query (#397)
We require two Postgres users, barman (a superuser) and streaming_barman
(a non-superuser with replication privileges). We connect as user barman
to execute queries (e.g., read pg_settings, or run pg_start_backup), and
as streaming_barman to create a replication slot and run pg_receivewal.

In Postgres v10+, the barman user does not need to be a superuser, but
can be a regular user that is granted permissions to read settings and
execute backup functions. Commit #8f5a5af added support for this setup,
but incorrectly introduced "userepl" privileges into the equation.

The barman user does not need the "replication" privilege, which grants
neither EXECUTE permissions on backup functions, nor the ability to read
settings (i.e., doing `createuser --replication barman` would still mean
having to GRANT all required privileges separately).

As reported in #318 and #319, this confusion results in some unfortunate
behaviour (a failing check and an empty exception). We fix both problems
here by removing userepl from the has_backup_privileges query, and also
introducing a better exception message.

Now if the barman user is not a superuser, it must be a member of the
pg_read_all_{stats,settings} roles (or a member of pg_monitor, which
implies both), and it must also have EXECUTE privileges on functions
pg_{start,stop}_backup, pg_create_restore_point, and pg_switch_wal.
The requirements for streaming_barman are unchanged; it must have
replication privileges.

Closes #318
Closes #319