Fix recovery_window_report test
Fixes the recovery_window_report test which was passing for the wrong
reason and not actually testing for the correct behaviour.

Firstly, each entry in `backup_source` is no longer referencing the same
`LocalBackupInfo` object which means `backup_source` now contains three
`LocalBackupInfo` objects with different values of end_time.

This initial fix unearthed a second issue which is that the assertion
of backup validity was incorrect: `test_backup2` is actually considered
VALID because it is required to recover to any point in time during the
recovery window before the begin_time of test_backup3.

There was a further bug in the assertion of the log content which was
passing even though the expected message wasn't found because the find
function returns `-1` if nothing is found, and `-1` is a valid
assertion.

Fixing this and changing the expected backup ID in the log message
unearthed a final issue which is that the value of
`_point_of_recoverability` is calculated from `datetime.now` which
means the value returned in the test is different to that returned while
the code-under-test is generating the retention report.

We work around this by using a regex match instead and avoid matching
the timestamp.