Do not overwrite begin_time when reading backup_label
Check whether a `backup_info` already has `begin_time` before setting it
from a backup label in `_backup_info_from_backup_label`.

This resolves a discrepancy between the `begin_time` and `end_time`
timezones in the backup info because both `begin_time` and `end_time`
are now timestamps obtained directly from PostgreSQL and therefore have
the same timezone as the PostgreSQL server.

This only affects the `PostgresBackupStrategy` which still needs to read
the backup label when backing up a server in recovery mode as it cannot
get the WAL name and offsets directly from PostgreSQL in such cases.

The only other backup strategy which calls
`_backup_info_from_backup_label` is `ConcurrentBackupStrategy` when
backing up a <9.6 server. In these cases it uses pgespresso and
`begin_time` is not set by the time `_backup_info_from_backup_label`
is called.

Closes #206