Very quick and cheap POC for IBM COS
In principle this is pretty straightforward - the IBM COS becomes a
subclass of the S3 cloud interface and deals with IBM creds there.

We need to abstract the actual boto3 module choice away from the
aws_s3 module though because it's now just referencing the IBM
one, which will likely not work against AWS (and even if it does work,
using the IBM fork against real S3 is not a great idea).

Testing is a bit of a pain - in theory the only new logic is really
the IBM auth, but if we're suppoint ibm_boto3 against IBM COS then
we'd want our integration tests to validate this works. We can take
the same approach we use for AWS and run the barman-cloud commands
with --cloud-provider=ibm-cos against minio, which at least validates
that S3CloudInterface works using ibm_boto3 against an S3-compatible
service (minio).