Handle EnterpriseDB version strings
Adds a special case in `PostgreSQLConnection.server_txt_version` to
handle the result of `SELECT version();` when it returns the
EnterpriseDB version instead of the PostgreSQL version - this is the
case in EPAS 9.6 and EPAS 10. More recent versions of EPAS return the
PostgreSQL version followed by the EnterpriseDB version so are already
handled correctly.

This means that we determine the same major version when running against
EPAS 9.6 that we do when running against PostgreSQL 9.6 (and likewise for
EPAS 10 and PostgreSQL 10) which means that the `--include` patterns
passed to RSync will match the tablespace directories.

This fixes a bug where tablespaces were not included in backups when
running against EPAS versions 9.6 or 10.

Note: An alternative fix for this would be to leave server_txt_version
alone and instead update utils.simplify_version so that it removes the
last field then returns the first two (or fewer) elements, specifically
we would change `version[:-1]` to `version[:-1][:2]`. While this would
do the right thing for all current version strings it is not clear at
all from the code why this is the right thing - it is just making more
assumptions about versions which happen to hold for all known supported
versions at this point in time.

By handling the "EnterpriseDB" case explicitly in server_txt_version we
are removing the additional version field only when we expect it to be
there and ensuring the response always matches the PostgreSQL version.
The existing code to convert the PostgreSQL version to a major version
can then continue to apply the "always remove last field" logic.

Closes #424