Fix two config-mutation problems
Currently, Barman operations are self-contained, loading its configs
each time something is run. As a result, the config-related code does
things like clobber the simple int or string loaded from the config file
with an object reference, and assumes that at such points in the code,
it will only be dealing with the initial types/values and never the
updated ones.

When this code is rerun in the same process (as it will have to in order
to support the backup REST API), these assumptions cause crashes.