Add test for cli help output
Adds a test which verfies the help output of the ArgumentParser used by
`cli.py`. Rather than test the actual parser created by `cli.py` we
instead create a minimal parser and check its help output against an
expected output. This removes any dependency between the actual content
of the `cli.py` parser so we do not have to update this test every time
we change a command or argument in `cli.py`.

The rationale for this test is: We depend on implementation details of
argparse in order to achieve this ordering and, while argparse has
remained stable between python 2.7 and 3.10, we must assume things we
depend on will change in the future. When that happens this test will
fail and we'll be able to update barman before anything broken gets
shipped to users.