[wip] Retire last_position from sync_info
last_position is an optimisation which makes it possible to avoid scanning
the entire contents of xlog.db however the trade-off is that it exposes an
implementation detail of xlog.db (that it uses byte offsets).

By removing last_position we can continue retiring xlog.db without losing
any barman functionality. The cost is that we always need to scan the
WAL archive. This is something which storage tiers themselves should be
optimising though.