Update barman-cloud exit status codes #438
Following the change in barman-cloud-wal-restore where the exit code for
network related errors is changed from 1 to 2, this commit updates the
remaining barman-cloud commands so that connectivity errors return an
exit status of 2.

We also change the exit status for any unhandled exception from 1 to 4,
and the exit status for bad input or precondition failures (e.g. a target
directory which already has content) is changed from 1 to 3.

This means the exit status of 1 can now represent the case where
the operation failed to complete even though there are no issues with
connectivity and command input.

Relates to #438.