Explicitly handle bad retention policy exceptions #438
Updates barman-cloud-delete so that it uses CLIErrorExit to exit with
status 3 in the event that a requested retention policy is not valid.

Rather than rely on ValueError, which may be raised by other code
called as part of `RetentionPolicyFactory.create`, we create a specific
exception just for bad retention policies and raise that instead. This
means the code within Barman (not just Barman Cloud) is also updated to
catch this more specific exception.

Relates to #438.