Add tags to backups and WALs uploaded to the cloud #476
Allows tags to be added to backups and WALs uploaded to cloud providers
using barman-cloud-backup and barman-cloud-wal-archive.

Both commands gain a new argument: --tags. The barman-cloud-wal-archive
command also gains a --history-tags argument which allows additional
tags to be specified which are added to history files instead of any
tags provided in the --tags argument.

Both --tags and --history-tags use the nargs="*" form, e.g.:

    --tags k1,v1 k2,v2 --history-tags k3,v3 k4,v4

Each argument is parsed with csv.reader using the `,` delimiter and the
default quote char of `"`.

If supplied, the tags will be passed to the cloud provider when
uploading WALs or backup files.

The implementation is achieved by adding a tags attribute to the cloud
interface which the implementations can then read and add via the
specific cloud provider library (the exact details of how and where tags
are added will vary between providers).

The tags attribute is read when uploading files via the upload_fileobj
method and the multipart streaming method. To handle --history-tags, an
override_tags option is added to upload_fileobj which allows the caller
to pass in a different set of tags to be used instead.

Note that the Azure and AWS implementations add the tags for multipart
uploads in different places: For AWS the tags are added when a multipart
request is created, for Azure the tags are added when a multipart request
is completed.

Closes #476.