Expose Azure client concurrency parameters #472
Adds options to the barman-cloud-wal-archive command which can be used
to control the concurrency of WAL upload operations. The options are:

- max_single_put_size: The maximum size of an object before the Azure
  client will use the concurrent chunk upload method. If this is reduced
  from the default of 64MB to a value below the size of the archived
  WAL files then the Azure client will split a single WAL file into
  chunks which are uploaded concurrently. This can reduce the overall
  WAL archive time.
- max_block_size: The maximum size of a chunk when using the concurrent
  chunk upload method.
- max_concurrency: The maximum number of chunks which can be uploaded
  concurrently.

The default values in the CLI have been chosen to preserve the existing
behaviour which is that any WAL with a post-compression size of <64MB
will be uploaded in a single request with no chunking.

The default values in AzureCloudInterface are deliberately different.
These are set to values which have been demonstrated to reduce WAL
archive times between Azure regions for WAL files with the default
PostgreSQL size of 16MB.

The reason the CLI defaults are set to maintain the current behaviour is
so that users can opt-in to the concurrent chunked upload method. In the
future we may change the CLI defaults to match the AzureCloudInterface
defaults.

Relates to #472.