Increase maximum Azure connection pool size #472
Increases the maximum connection pool size on the requests session used
by the Azure client from the default of 10 to 32. This allows up to 32
concurrent connections to be used when archiving WAL files to Azure
Blob Storage. Whether all 32 connections are used depends on the values
of max_single_put_size, max_concurrency and max_block_size.

Relates to #472.