Add snappy compression option to Barman Cloud #344
Adds the `--snappy` option to barman-cloud-backup and
barman-cloud-wal-archive which causes Barman Cloud to apply the snappy
compression algorithm to the data uploaded to the cloud provider. This
yields faster performance but poorer compression ratios than gzip and
bzip2.

As with bzip2 and gzip compression, the barman-cloud-restore and
barman-cloud-wal-restore commands will automatically detect when snappy
compression has been applied and decompress using snappy while
downloading.

In terms of implementation, the compression-from-file-to-cloud and
decompression-from-cloud-to-file cases (used for WAL archive/restore as
well as uploading backup.info during a backup) are reasonably trivial.
The snappy library is used to compress and decompress between in-memory
file-like objects in a very similar way to the built-in gzip and bz2
libraries. The implementation can be found in the compress and
decompress_to_file functions of the new barman.clients.cloud_compression
module, along with the analogs for gzip and bz2.

Compression/decompression for backup and restore is a little more
complicated due to the way the backup is compressed/decompressed in
chunks on-the-fly. Currently this is handled by the Python TarFile
implementation which has built-in support for gzip and bz2.

For compression during a backup, CloudTarUploader is updated so that it
will create a reference to a custom compressor if a compression other
than gzip or bz2 is requested. This custom compressor is used in the
write function (called by TarFile) to compress the bytes before writing
them to the buffer which is, eventually, uploaded to cloud storage via
multipart requests (or equivalent, if a cloud provider does not directly
support multipart requests).

For decompression during a restore we need to decompress the bytes before
they are passed to TarFile.extractall in CloudInterface.extract_tar
while avoiding reading the entire archive into memory. For this we use
composition to add a second chunked downloader which reads the chunks
from the RawIOBase subclass used to download the compressed chunks from
the cloud provider. This is implemented in
barman.cloud.DecompressingStreamingIO which will read compressed bytes
from any RawIOBase subclass (this will be either a StreamingBodyIO for
S3 or a StreamingBlobIO for Azure Blob Storage), decompress them and
return the requested number of bytes, storing any left over in a local
buffer which is used to serve subsequent reads.

Closes #344.