Make python-snappy dependency optional
Moves the python-snappy dependency to extras_require and lazily
imports it.

This is required so that we do not break barman on platforms which do
not have a sufficiently recent python-snappy package available in their
OS packages.

Users who wish to use --snappy compression will need to install
python-snappy with pip, the same way the cloud dependencies boto3,
azure-identity and azure-storage-blob are installed.

In the future we should look at resolving these dependencies at the OS
package level so that barman cloud does not require the manual
installation of any additional python packages.