Use calculated timeline when checking whether to PITR
Replaces `target_tli` with `calculated_tli` when assigning the value
of `d_tli`.

The value `d_tli` is used to determine whether the requested target
timeline should cause Barman to prepare for PITR. The logic Barman
applies is that the target timeline should only require PITR if
it differs from the timeline which was current when the backup was
taken. Because `target_tli` may be `current` or `latest` we cannot
compare it to `backup_info.timeline` (which is always numeric)
however we can use `calculated_tli` instead.

Relates to #305.