Avoid copying contents of pg_tblspc to passive server
Adds an entry to to the exclude_and_protect filter for each tablespace
OID in `/data/pg_tblspc` so that any symlinks present are not
transformed into their referent directories by rsync during the sync of
the backup to a passive Barman server.

This is required in order to stop rsync from tranforming such symlinks
into their referent directories and copying them onto the passive server
along with the link targets in the root of the base backup directory.
This would result in the tablespace data being duplicated on the passive
server.

Such symlinks are only created on the primary Barman server when using
`backup_method=postgres`. When `backup_method=rsync` is used Barman
excludes the symlinks in `pg_tblspc` and therefore they are not present
when syncing to the passive server.

As a result of this patch the symlinks created during a backup with
`backup_method=postgres` will not be copied to the passive server at all
however this is not a problem because Barman will recreate the symlinks
itself during a `barman recover` using the information in `backup.info`.

Closes #309 and #516.

Thanks to @jfung-github and @hzetters for reporting the bug.