Add compression option to `barman backup`
Adds support for pg_basebackup compression to the `barman backup`
command.

The following new global/server config variables are added:

  - backup_compression: defines the compression algorithm to use
    (currently only gzip is supported however this will change with
    PostgreSQL 15).
  - backup_compression_level: defines the level of compression to be
    applied.

The scheme `backup_compression_*` was chosen to avoid conflicting with
the existing `compression` configuration variable which concerns WAL
compression.

A new class is added to the compression module to manage pg_basebackup
compression. A factory method to instantiate the class for a given
compression method is also added.

This class contains the details of how such compression options are
validated as well as how files compressed with that compression are
opened. This provides (almost) a single point in the code where support
for a new pg_basebackup compression algorithm would be added.

Other than adding the pg_basebackup arguments required by the supplied
options, the functional changes are:

  - Add the compression type to the BackupInfo.
  - Read the backup_label from the compressed tarballs and add to the
    BackupInfo.