[1/?] Add compression option to `barman backup`
Adds the `--compression` option to the `barman backup` arguments and
also adds global/config variables for the same. The compression type
to be applied by pg_basebackup can be set using this option and it will
be stored in `server.config.backup_compression`.

The name `backup_compression` was chosen to avoid conflicting with the
existing `compression` configuration variable which concerns WAL
compression.