[2/?] Use compression args with pg_basebackup
Updates the backup_executor so that it passes the backup_compression
for the server to the pg_basebackup command and updates the command
wrapper so that it creates the pg_basebackup options needed to yield
the compression requested by backup_compression.

Although backup_compression is stored in the server config, the
BackupExecutor config is a reference to the BackupManager config which
is a reference to the server config, so we can just reference the
backup_compression attribute of the BackupExecutor config directly.