[3/?] Add `--compression-level` option to barman backup
Adds the `--compression-level` option to barman backup which currently
maps to the `--compress` pg_basebackup option.

We do not place any restriction on what can be passed in other than
it must be a positive integer. Different compression types support
different levels so we will just pass whatever the user gives us for
now.