[4/?] Read backup label from compressed backups
Updates the BackupStrategy so that the backup label can be read from
a compressed backup if the backup is compressed.

So that the BackupStrategy knows whether to retrieve the label from
a compressed file or not we add a field to the backup.info in which
the compression type is recorded. This field will also be required
to recover such backups.

There are some caveats to this:

1. Reading from the tar is relying on the assumption that the backup
   label will be the first member. This is not documented behaviour
   so can't be relied on.
2. The function to read the backup label assumes gzip. For PG15 we
   will want to support other compressions too so this should be
   abstracted a little - perhaps a BackupLabelStrategy, or maybe
   just an if statement.