[wip] Start consolidating compression-specific details
Introduces a BackupCompressionManager (no real relation to the existing
CompressionManager other than the name...) which will encapsulate the
compression-specific knowledge so that we're not distributing it
throughout the codebase.

Initially it is used to provide a reference to the backup_compression
config property which automatically defaults to gzip if only
backup_compression_level is set.

This means the backup manager no longer needs to care about the detail
and can just reference the property.