Allow compression location (client/server) to be specified
Adds a new global/server option backup_compression_location which
can be used to determine whether compression takes place on the
server or the client.

Values of "client" are allowed for any pg_basebackup and PostgreSQL
version since the default pg_basebackup behaviour is to compress
on the client.

Values of "server" require >= 15 for both pg_basebackup and the
PostgreSQL server, since older pg_basebackup versions don't support
the option and older PostgreSQL servers don't have the server-side
compression feature.