Add backup_compression_format option
Adds a new global/server config option backup_compression_format.

This maps to the pg_basebackup --format option and behaves in the same
way:

- For compression on the client the format is always tar.
- For compression on the server the format can be tar or plain.
  If plain is used then the server will compress the backup for network
  transmission and the client will uncompress it before writing to disk.

The compression field in BackupInfo is now only set if the compression
format is tar. This prevents the backup.info file recording the
compression type if it was only used during network transmission.