Implement recovery for pg_basebackup compressed backups
Allows `barman recover` to recover backups compressed by pg_basebackup.
Recovery is implemented by using Rsync to copy the compressed files to
a staging directory on the recovery server and then extracting the
tarballs into the recovery destination using the tar command. The config
files are then copied back to the Barman server from the recovery
destination so that Barman can perform its analysis and mangling of the
configuration. Rsync is then used again to copy the mangled config files
backup to the recovery destination. Finally the staging directory on
the recovery host is deleted.

Retrieving the config files from the recovery destination is a different
approach to that used in recovery of a plain backup. In the plain backup
case, Barman has always copied the file from the base backup directory
in order to perform its analysis and mangling. This is a cheaper
operation than copying config files from the recovery destination since
it avoids copying the file over the network. With compressed backups
the config files are in compressed tarballs and these will need to be
scanned in order to extract them. For a sizeable backup such a scan
would take significantly longer than the time required to copy the
already-extracted file back to the Barman server. We therefore avoid
this potentially expensive operation and instead copy the extracted
configuration files backup to the Barman server.

We require that the user specifies the path in which the staging
directory is creaeted (note: this is not the path to the staging
directory but the path in which the staging directory will be created).
This can be specified using the `recovery_staging_path` config option
or the `--recovery-staging-path` CLI option. The reason the user must
supply it is because Barman cannot make any safe assumptions about
where the compressed backup data can be staged.

A suitable staging location would be:

  1. Only accessible to the postgres user.
  2. Mounted on local storage.
  3. Have sufficient space available to stage the compressed backup
     files.

Co-authored-by: Didier Michel <didier.michel@enterprisedb.com>