[wip] Fix config mangling for compressed backups
One of the final steps of the recovery process is to mangle the config
files.

This means:

1. Checking the presence of the config files in the backup.
2. Copying them to a temporary location where they can be edited.
3. Appending any required lines to postgresql.auto.conf (for PITR).
4. Identifying and commenting out dangerous options.
5. Warning about other dangerous options.
6. Copying the files from the temporary location into the remote
   recovery location.

For non-remote recoveries, steps 2 and 6 are not needed - the files are
mangled in place in the recovery directory.

The methods which mangle the config save the original with an `.origin`
suffix.

When we recover compressed backups this entire process fails because
the config files cannot be found in the backup directory (they are in
compressed tarballs).

This commit attempts to resolve this by:

1. Extracting the code in _map_temporary_config_files which checks for
   the existence of the config files and copies them into the temporary
   location. The code is now in two new methods - one which lists the
   files, one which copies them.
2. Overriding these methods in the TarballRecoveryExecutor subclass so
   that they methods are "tarball aware".

This means that by the time _map_temporary_config_files has finished
it has the correct list of config files and they are available in the
temporary directory where the following two methods (
_analyse_temporary_config_files and _copy_temporary_config_files) expect
to find them.

This implementation deliberately uses the tarballs in the staging
directory on the PG server as the source for the config files instead
of the tarballs in the backup directory on the Baman server. This is
because when encryption is added in the future we will not necessarily
be able to read the tarballs on the Barman server at all, whereas it is
reasonable to assume the tarballs will eventually be available on the
PG server in decrypted form.

The advantage of this approach is that we make minimal changes to the
codebase - we don't really want to touch the details of the config
mangling if we can avoid it.

There are some disadvantages:

1. We're having to pass variables to the new methods which they don't
   actually need, since the uncompressing and regular versions need
   different things (one needs a remote connection, the other needs to
   know where the backup directory is). This is partially a result of
   inheritance not being the best code reuse strategy in this case but
   may also hint at the extraction of the specific code being at the
   wrong level.
2. Copying the config files from the PG server to the Barman server
   only to then copy them back is a bit wasteful, though given the
   config files are tiny compared to the size of the backup the
   practical impact of this extra work is likely minimal.
3. Listing the files in the tarball before then extracting them is a bit
   wasteful. We might be better off merging that into one step so we
   just try to extract all of them and figure out what doesn't exist
   as a result.

There are also general flaws with this commit - lots of hardcoded
filenames and .tar.gz, sloppy code etc. Hopefully it illustrates the
general idea though and does demonstrate it works.