Do not wait for WAL archiving in PG backup stop function
Changes the value of the `wait_for_archive` argument to pg_backup_stop
(or pg_stop_backup for PG<15) calls from the default TRUE to an explicit
FALSE.

This changes the behaviour of `barman backup` such that Barman no
longer asks PostgreSQL to wait for WAL archiving during the stop backup
call and instead relies on its own wait process which happens later in
the backup.

Barman's own wait process is determined by the use of the `--wait` and
`--wait-timeout` options on the `barman backup` command. If `--wait` is
used then Barman will wait until the WALs required for consistency of
the backup have been archived *by Barman*. This means they have been
moved from the incoming or streaming directory into the wals directory
and a record has been written to `xlog.db`.

This change allows users to omit the `--wait` option in situations where
they do not want to wait for WAL archiving to complete before the backup
command finishes. A typical scenario is backup of a standby when there
is minimal traffic on the primary server. Prior to this commit, such
cases would mean `barman backup` would hang until a WAL switch occurred
on the primray, regardless of whether `--wait` was omitted or not.

With this change, the `--wait` option can be omitted *or* a suitable
timeout can be set with `--wait-timeout` and this will allow the backup
process to complete even when backing up a standby with minimal traffic
on the primary.

In such situations it is essential that archiving is monitored and that
`barman cron` is running so that the backup can eventually move from the
`WAITING_FOR_WALS` state to the `DONE` state.

It is important that users upgrading to Barman 2.20 (the first release
to include this change) are aware that Barman will now *only* wait for
the necessary WALs if the `--wait` option is included. Previously Barman
would ask PostgreSQL to wait regardless of whether `--wait` was used
which meant even if Barman hadn't archived the WALs, PostgreSQL was at
least guaranteed to have copied or streamed them onto the Barman server.
This is no longer the case. Unless the `--wait` flag is used the
`barman backup` command will complete successfully regardless of whether
the WALs required for backup consistency have been archived.

This will only be a problem for users who are:

* Only relying on the result of `barman backup` to determine backup
  success and not monitoring backup state (i.e. checking that backups
  end up in the DONE state), and:
* Not monitoring WAL archiving / streaming, and:
* Not using the `--wait` option on `barman backup`.

This should be highlighted in the release notes.

Closes #448.