Set current_action at start of stop_backup method
Moves the line in ConcurrentBackupStrategy.stop_backup which sets the
current_action to the very beginning of the function.

This resolves an issue where a failure retrieving the server_version
from the PostgreSQL connection would result in a misleading error
message due to the current_action still being set to
"issuing start backup command".

Closes #496