Fix --immediate-checkpoint backup argument
Updates the handling of the immediate_checkpoint argument so that it
sets the immediate_checkpoint property of the postgres connection in
addition to the server config.

This fixes a bug where the `--immediate-checkpoint` argument would be
ignored for Rsync backups because it was only updated in the server
config. Because the server's postgres connection has already been
created from the config by this point it means its immediate_checkpoint
attribute will remain set according to the config, preventing the value
supplied in the argument from having any effect.

This did not affect backups taken with `backup_method = postgres`
because the immediate_checkpoint argument is read from the config when
creating the PgBaseBackup command wrapper.

Closes #561