Treat timelines as hex when inferring from WAL segment
Parses the timeline as a base 16 integer when resolving the special
value `latest` for the `--target-tli` option. This change is made in
Server.get_required_xlog_files and RecoveryExecutor._set_pitr_targets
which are the two places where the timeline is read from the WAL archive
and converted to a numeric ID.

This fixes a bug where `--target-tli=latest` would not resolve to the
correct timeline when the latest timeline was greater than 9.

Closes #581.

Thanks to Amit Singh for finding this bug.