Do not assume custom compression when magic is set
Updates the CompressionManager such that unidentified compression is
no longer assumed to be custom compression *if* the magic bytes are set
in `config.custom_compression_magic`. If the magic bytes are not set
then the CompressionManager will continue to assume that unidentified
compression is custom compression.

This allows users who set custom_compression_magic to avoid an issue
when `barman-wal-restore` is used to fetch uncompressed WALs directly
from the `streaming` directory during recovery. Such WALs would
previously fail compression identification and so Barman would try to
uncompress them using the custom compression. This situation will now
be avoided (as long as custom_compression_magic is set).

If custom_compression_magic is not set then the original assumption is
preserved and Barman will try and uncompress any WALs which it cannot
identify using the custom compression. This behaviour is preserved for
backward compatibility but should be dropped in a major version release
such that users are required to specify custom_compresion_magic (and
therefore are forced to avoid the issue described above).

Closes #566.

Thanks to dsmoljanovic for the very clear and detailed bug report.